changed update modal to icon in todo list
