+++ client/src/components/TodoList.js
@@ -4,16 +4,29 @@ import * as Icon from 'react-bootstrap-icons';
 import { AuthContext } from '../providers/AuthProvider';
 import CreateTodoInput from './CreateTodoInput';
 import ShowTodoItem from './ShowTodoItem';
+import UpdateNoteModal from './UpdateNoteModal';
 
 const TodoList = (props) => {
   const { user } = useContext(AuthContext)
-  const { note } = props
+  const { note, displayNote, getTodoNotes, getCompletedNotes } = props
   const [items, setItems] = useState(null)
   const [hideTrash, setHideTrash] = useState(false)
+  const [ updateNote, setUpdateNote ] = useState(null)
+
+
 
   useEffect(()=>{
     getItems()
   },[])
+
+  
+  const openUpdateNoteModal = (note) => {
+    setUpdateNote(note)
+  }
+
+  const closeUpdateNoteModal = () => {
+    setUpdateNote(null)
+  }
 
   const getItems = async () => {
     try {
@@ -90,8 +103,15 @@ const TodoList = (props) => {
 
   return(
     <>
-      <h1>{note.title}</h1>
-      <h5>{note.body}</h5>
+      <div>
+        <div>
+          <h1>{note.title}</h1>
+          <h5>{note.body}</h5>
+        </div>
+        <div style={{cursor: 'pointer'}} onClick={() => openUpdateNoteModal(note)}>
+          <Icon.PencilSquare />
+        </div>
+      </div>
       <div className='todo-list'>
         <div className='todo-list-title'>
           <h6>Todo</h6>
@@ -105,6 +125,13 @@ const TodoList = (props) => {
         <h6>Completed</h6>
         {items && renderCompletedItems()}
       </div>
+
+      {updateNote && 
+        <UpdateNoteModal
+          closeUpdateNoteModal={closeUpdateNoteModal} updateNote={updateNote} 
+          getTodoNotes={getTodoNotes} getCompletedNotes={getCompletedNotes}
+          displayNote={displayNote} note={note} getItems={getItems}
+      />}
     </>
   )
 }
